# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  swagmodel.GetRating:
    properties:
      articleID:
        type: integer
      id:
        type: integer
      star:
        type: integer
    type: object
  swagmodel.GetRatingSum:
    properties:
      articleID:
        type: integer
      avgStar:
        type: integer
      id:
        type: integer
    type: object
  swagmodel.GetUser:
    properties:
      avatarURL:
        type: string
      bio:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  swagmodel.GetUserPublic:
    properties:
      avatarURL:
        type: string
      bio:
        type: string
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
...
  swagmodel.InputRating:
    properties:
      star:
        type: integer
    type: object
  swagmodel.InputUser:
    properties:
      avatarURL:
        type: string
      bio:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
  swagmodel.Response:
    properties:
      msg:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: hdkef11@gmail.com
    name: Hadekha Erfadila Fitra
  description: Bloggo is a REST Api for blogging
  title: Bloggo API
  version: "1.0"
paths:
...
  /ratings/{article-id}:
    get:
      description: get all ratings navigated by last-id and limit.
      parameters:
      - description: ID of article
        in: path
        name: article-id
        required: true
        type: integer
      - description: ID of the last rating in recent array of rating
        in: query
        name: last-id
        required: true
        type: integer
      - description: how many rating you want to take
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/swagmodel.GetRating'
            type: array
      summary: get all ratings
      tags:
      - ratings
    post:
      consumes:
      - application/json
      description: create an rating
      parameters:
      - description: ID of article
        in: path
        name: article-id
        required: true
        type: integer
      - description: create an rating
        in: body
        name: rating
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputRating'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: create an rating
      tags:
      - ratings
  /ratings/{article-id}/sum:
    get:
      description: get summed ratings navigated by last-id and limit.
      parameters:
      - description: ID of article
        in: path
        name: article-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.GetRatingSum'
      summary: get summed ratings
      tags:
      - ratings
  /ratings/{id}:
    delete:
      consumes:
      - application/json
      description: delete an rating
      parameters:
      - description: ID of rating
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: delete an rating
      tags:
      - ratings
    put:
      consumes:
      - application/json
      description: update an rating
      parameters:
      - description: ID of rating
        in: path
        name: id
        required: true
        type: integer
      - description: update an rating
        in: body
        name: rating
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputRating'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: update an rating
      tags:
      - ratings
...
  /users:
    get:
      description: get all users navigated by last-id and limit.
      parameters:
      - description: ID of the last user in recent array of user
        in: query
        name: last-id
        required: true
        type: integer
      - description: how many user you want to take
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/swagmodel.GetUser'
            type: array
      security:
      - AuthToken: []
      summary: get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: create a user (BY SUPER-ADMIN)
      parameters:
      - description: create an user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: create a user (BY SUPER-ADMIN)
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: delete an user
      parameters:
      - description: ID of user
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: delete an user
      tags:
      - users
    get:
      description: get detail of user specified by id
      parameters:
      - description: ID of user
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.GetUser'
      security:
      - AuthToken: []
      summary: get one user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: update an user
      parameters:
      - description: ID of user
        in: path
        name: id
        required: true
        type: integer
      - description: update an user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: update an user, only super-admin can update role
      tags:
      - users
  /users/{id}/public:
    get:
      description: get detail of user specified by id
      parameters:
      - description: ID of user
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.GetUserPublic'
      summary: get one user public information (without password information)
      tags:
      - users
securityDefinitions:
  AuthToken:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"
...